+++ packages/mobile/ios/NodeJsMobile/libnode/include/node/openssl/objects.h
@@ -1,20 +1,14 @@
 /*
- * Copyright 1995-2019 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 1995-2018 The OpenSSL Project Authors. All Rights Reserved.
  *
- * Licensed under the Apache License 2.0 (the "License").  You may not use
+ * Licensed under the OpenSSL license (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
  * in the file LICENSE in the source distribution or at
  * https://www.openssl.org/source/license.html
  */
 
-#ifndef OPENSSL_OBJECTS_H
-# define OPENSSL_OBJECTS_H
-# pragma once
-
-# include <openssl/macros.h>
-# ifndef OPENSSL_NO_DEPRECATED_3_0
-#  define HEADER_OBJECTS_H
-# endif
+#ifndef HEADER_OBJECTS_H
+# define HEADER_OBJECTS_H
 
 # include <openssl/obj_mac.h>
 # include <openssl/bio.h>
@@ -26,9 +20,7 @@
 # define OBJ_NAME_TYPE_CIPHER_METH       0x02
 # define OBJ_NAME_TYPE_PKEY_METH         0x03
 # define OBJ_NAME_TYPE_COMP_METH         0x04
-# define OBJ_NAME_TYPE_MAC_METH          0x05
-# define OBJ_NAME_TYPE_KDF_METH          0x06
-# define OBJ_NAME_TYPE_NUM               0x07
+# define OBJ_NAME_TYPE_NUM               0x05
 
 # define OBJ_NAME_ALIAS                  0x8000
 
@@ -63,7 +55,7 @@ void OBJ_NAME_do_all_sorted(int type,
                             void (*fn) (const OBJ_NAME *, void *arg),
                             void *arg);
 
-DECLARE_ASN1_DUP_FUNCTION_name(ASN1_OBJECT, OBJ)
+ASN1_OBJECT *OBJ_dup(const ASN1_OBJECT *o);
 ASN1_OBJECT *OBJ_nid2obj(int n);
 const char *OBJ_nid2ln(int n);
 const char *OBJ_nid2sn(int n);
@@ -163,7 +155,7 @@ const void *OBJ_bsearch_ex_(const void *key, const void *base, int num,
 int OBJ_new_nid(int num);
 int OBJ_add_object(const ASN1_OBJECT *obj);
 int OBJ_create(const char *oid, const char *sn, const char *ln);
-#ifndef OPENSSL_NO_DEPRECATED_1_1_0
+#if OPENSSL_API_COMPAT < 0x10100000L
 # define OBJ_cleanup() while(0) continue
 #endif
 int OBJ_create_objects(BIO *in);